requirements.txt
Django==5.2.4
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
Pillow==12.1.1
gunicorn==23.0.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
white-noise==6.8.2
